requirements.txt
s3plz
flask
unidecode
slugify 
pyyaml
pytaglib
elasticsearch
gevent
requests
pytz
iso8601
tzlocal
